requirements.txt
pillow==10.2.0
imagehash==4.3.1
opencv-python==4.10.0.84
numpy==2.2.1